Refactoring: extract Commits to a separate file
