+++ dvc/git_wrapper.py
@@ -5,7 +5,8 @@ from dvc.config import Config
 from dvc.executor import Executor, ExecutorError
 from dvc.path.data_item import DataItemError
 from dvc.system import System
-from dvc.workflow import Workflow, Commit
+from dvc.graph.workflow import Workflow
+from dvc.graph.commit import Commit
 
 
 class GitWrapperI(object):
